consolidate register user functions inside user service
